integrate to IRONMAIDEN
zimbraMtaMaxMessageSize=0 being treated by ZWC as a 0 limit on attachment size rather than meaning "no limit" as it means for other parts of Zimbra.
GetInfoResponse attSizeLimit="-1" now treated as meaning "no limit".

https://reviewboard.eng.vmware.com/r/420902/

http://bugzilla.zimbra.com/show_bug.cgi?id=79003

Copied from Perforce
 Change: 437438

+++ src/java/com/zimbra/cs/service/FileUploadServlet.java
@@ -604,7 +604,7 @@ public class FileUploadServlet extends ZimbraServlet {
         try {
             // write the upload to disk, but make sure not to exceed the permitted max upload size
             long size = ByteUtil.copy(req.getInputStream(), false, fi.getOutputStream(), true, upload.getSizeMax() * 3);
-            if (size > upload.getSizeMax()) {
+            if ((upload.getSizeMax() >= 0 /* -1 would mean "no limit" */) && (size > upload.getSizeMax())) {
                 mLog.debug("handlePlainUpload(): deleting %s", fi);
                 fi.delete();
                 mLog.info("Exceeded maximum upload size of " + upload.getSizeMax() + " bytes: " + acct.getId());
